refactor: Separate each classes to each files
